requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
argo-workflows==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.6.3
bleach==5.0.1
blinker==1.7.0
bokeh==2.4.3
boto3==1.26.76
botocore==1.29.80
cachetools==5.2.0
certifi==2022.9.24
certipy==0.1.3
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==38.0.3
csscompressor==0.9.5
cycler==0.11.0
daiquiri==3.2.1
dask==2023.2.1
datashader==0.14.2
datashape==0.5.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
delegator.py==0.1.1
dictdiffer==0.9.0
distlib==0.3.6
distributed==2023.2.1
distro==1.8.0
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
Flask==3.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
greenlet==2.0.1
HeapDict==1.0.1
holoviews==1.15.1
hvplot==0.8.1
idna==3.4
imageio==2.22.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
invectio==0.2.2
ipykernel==6.17.1
ipynbname==2021.3.2
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonformatter==0.3.2
jsonschema==4.17.0
jupyter-highlight-selected-word==0.2.0
jupyter-kernel-gateway==2.4.0
jupyter-nbutils==0.1.3
jupyter-require==0.6.1
jupyter-server==1.23.2
jupyter-server-mathjax==0.2.6
jupyter-server-proxy==3.2.2
jupyter-telemetry==0.1.0
jupyter_client==7.4.7
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.0.0
jupyter_latex_envs==1.4.6
jupyter_nbextensions_configurator==0.6.1
jupyterhub==1.5.0
jupyterlab==3.1.19
jupyterlab-git==0.39.3
jupyterlab-pygments==0.2.2
jupyterlab-requirements==0.16.2
jupyterlab-spellchecker==0.7.2
jupyterlab-tabnine==0.0.24
jupyterlab_server==2.16.3
kiwisolver==1.4.4
kubernetes==11.0.0
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
micropipenv==1.4.4
mistune==2.0.4
mock==4.0.3
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbdime==3.1.1
nbformat==5.7.0
nbgitpuller==1.1.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.4.12
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
openshift==0.11.0
packaging==21.3
pamela==1.0.0
pandas==1.5.0
pandocfilters==1.5.0
panel==0.14.0
param==1.12.3
parso==0.7.1
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2020.11.15
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
plotly==5.13.1
pluggy==1.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyct==0.5.0
pyelftools==0.29
Pygments==2.13.0
pylab-sdk==1.3.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-string-utils==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc5424-logging-handler==1.4.3
rich==12.6.0
rich-click==1.5.2
rise==5.7.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.1
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.11.0
simpervisor==0.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
SQLAlchemy==1.4.44
supervisor==4.1.0
tblib==1.7.0
tenacity==8.2.2
termcolor==2.1.0
termcolor-whl==1.1.2
terminado==0.17.0
thamos==1.29.1
thoth-analyzer==0.1.8
thoth-common==0.36.6
thoth-python==0.16.11
threadpoolctl==3.1.0
tifffile==2023.2.27
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.8.0
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.1
xarray==2023.1.0
yarl==1.8.1
yaspin==2.2.0
zict==2.2.0
zipp==3.10.0
